Extract companionwindow selectors to a new file, and create some base selectors to set up static data structures that can get reused
