+++ src/state/selectors/index.js
@@ -1,5 +1,6 @@
 export * from './annotations';
 export * from './canvases';
+export * from './companionWindows';
 export * from './config';
 export * from './manifests';
 export * from './windows';
